function_app/requirements.txt
azure-functions
azure-storage-blob
azure-identity
httpx
